Fixed "can't create Thread" and "failed to allocate memory" errors on Windows.
When downloading cookbooks from a chef api location, berkshelf was
sometimes blowing up with "can't create Thread" or "failed to allocate
memory" errors.  This happened especially easily on Windows systems.
The error was due to a "thread leak" in chef_api_location.rb.  This leak
becomes apparent very easily on a Windows system when a berks install
downloads 20 to 25 (or so) cookbooks from a chef server (by which time
close to 1000 threads exist in the process).  This problem also exists
on linux, but generally a process on linux can create many more threads
before a problem shows up.