In chef_api_location finalizer, call terminate! instead of terminate (async instead of sync).
When running the tests, errors were occurring in Celluloid saying
"Couldn't cleanly terminate all actors in 10 seconds!"  (The tests call
Celluloid.shutdown before each example, which terminates all Actors, and
that shutdown call was producing this message).  The cause was that the
finalizer on the ChefAPILocation class was calling terminate on the
ridley connection synchronously.